--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -657,20 +657,13 @@
     - method
     - function
   return: argument 0
-  options:
-    - cname: ltValue
-      arguments:
-        - arg: THBoolTensor* result
-          output: True
-        - THTensor* self
-        - real other
-    - cname: ltTensor
-      arguments:
-        - arg: THBoolTensor* result
-          output: True
-        - arg: THTensor* self
-          broadcast: other fallback
-        - THTensor* other
+  cname: ltTensor
+  arguments:
+    - arg: THBoolTensor* result
+      output: True
+    - arg: THTensor* self
+      broadcast: other fallback
+    - THTensor* other
 ]]
 [[
   name: lt_
@@ -694,20 +687,13 @@
     - method
     - function
   return: argument 0
-  options:
-    - cname: gtValue
-      arguments:
-        - arg: THBoolTensor* result
-          output: True
-        - THTensor* self
-        - real other
-    - cname: gtTensor
-      arguments:
-        - arg: THBoolTensor* result
-          output: True
-        - arg: THTensor* self
-          broadcast: other fallback
-        - THTensor* other
+  cname: gtTensor
+  arguments:
+    - arg: THBoolTensor* result
+      output: True
+    - arg: THTensor* self
+      broadcast: other fallback
+    - THTensor* other
 ]]
 [[
   name: gt_