--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -652,20 +652,6 @@
         - THTensor* other
 ]]
 [[
-  name: lt
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: ltTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
-]]
-[[
   name: lt_
   return: self
   options:
@@ -680,20 +666,6 @@
         - arg: THTensor* self
           broadcast: other inplace fallback
         - arg: THTensor* other
-]]
-[[
-  name: gt
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: gtTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
 ]]
 [[
   name: gt_
@@ -712,20 +684,6 @@
         - THTensor* other
 ]]
 [[
-  name: le
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: leTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
-]]
-[[
   name: le_
   return: self
   options:
@@ -740,20 +698,6 @@
         - arg: THTensor* self
           broadcast: other inplace fallback
         - THTensor* other
-]]
-[[
-  name: ge
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: geTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
 ]]
 [[
   name: ge_
@@ -772,20 +716,6 @@
         - THTensor* other
 ]]
 [[
-  name: eq
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: eqTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
-]]
-[[
   name: eq_
   return: self
   options:
@@ -800,20 +730,6 @@
         - arg: THTensor* self
           broadcast: other inplace fallback
         - THTensor* other
-]]
-[[
-  name: ne
-  variants:
-    - method
-    - function
-  return: argument 0
-  cname: neTensor
-  arguments:
-    - arg: THBoolTensor* result
-      output: True
-    - arg: THTensor* self
-      broadcast: other fallback
-    - THTensor* other
 ]]
 [[
   name: ne_