--- a/aten/src/ATen/Dispatch.h
+++ b/aten/src/ATen/Dispatch.h
@@ -69,3 +69,21 @@
             "%s not implemented for '%s'", (NAME), the_type.toString()); \
     }                                                                    \
   }()
+
+#define AT_DISPATCH_ALL_MATH_TYPES(TYPE, NAME, ...)                      \
+  [&] {                                                                  \
+    const at::Type& the_type = TYPE;                                     \
+    switch (the_type.scalarType()) {                                     \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Byte, uint8_t, __VA_ARGS__)   \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Char, uint8_t, __VA_ARGS__)   \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Double, double, __VA_ARGS__)  \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Float, float, __VA_ARGS__)    \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Int, int32_t, __VA_ARGS__)    \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Long, int64_t, __VA_ARGS__)   \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Short, int16_t, __VA_ARGS__)  \
+      AT_PRIVATE_CASE_TYPE(at::ScalarType::Half, half, __VA_ARGS__)      \
+      default:                                                           \
+        at::runtime_error(                                               \
+            "%s not implemented for '%s'", (NAME), the_type.toString()); \
+    }                                                                    \
+  }()