--- a/aten/src/THC/generated/THCTensorMathCompareByte.cu
+++ b/aten/src/THC/generated/THCTensorMathCompareByte.cu
@@ -1,3 +1,0 @@
-#include "../THCTensorMathCompare.cuh"
-#include "../generic/THCTensorMathCompare.cu"
-#include "../THCGenerateByteType.h"