Fixes #5008. Instead of throwing error, we allow doing comparators with floats on integral tensors.

example:
```
>>> a = torch.LongTensor([0,1]).cuda()
>>> a.ge(0.5)

 0
 1
[torch.cuda.ByteTensor of size (2,) (GPU 0)]
```