@zou3519 2018-02-24
Could you run some benchmarks for before/after numbers?