@li-roy 2018-03-01
Ran some benchmarks on ge for tensors of size (100, 100, 100). Seems like no perf regression.

new:
floats: 1.06 ms ± 11.7 µs
floats gpu: 43.6 µs ± 49.8 ns
longs: 971 µs ± 9.13 µs
longs gpu: 59 µs ± 40.3 ns

old:
floats: 1.06 ms ± 26 µs per loop
floats gpu: 43.6 µs ± 30.5 ns per loop
longs: 956 µs ± 6.38 µs per loop
longs gpu: 59 µs ± 31.2 ns per loop