@li-roy 2018-03-01
This doesn't compile on windows because of the use of std comparators. Is it possible to turn this flag on? Alternatively I could change the template parameter to take in a function instead.

```
00:53:31 C:/Jenkins/workspace/pytorch-builds/pytorch-win-ws2016-cuda9-cudnn7-py3-build/aten/src/ATen/native/cuda/TensorCompare.cu(70): error: calling a constexpr __host__ function("operator()") from a __device__ function("operator()") is not allowed. The experimental flag '--expt-relaxed-constexpr' can be used to allow this.

00:53:31           detected during instantiation of "void at::native::CmpOpFloatingCUDA<Comparator>::apply(at::Tensor &, const at::Tensor &, at::Scalar) [with Comparator=std::less_equal]" 

00:53:31 (84): here
```