@apaszke 2018-03-01
Can you put the benchmark script somewhere online? I'm not sure what the values mean