@ezyang 2018-03-06
@pytorchbot retest this please