@li-roy 2018-03-08
Here's some benchmarks I ran. I only did ge, but all others are similar.
benchmark ipython script: https://gist.github.com/li-roy/76046c3dae52dfb93d376ffd1187836c
this branch results: https://gist.github.com/li-roy/58239f19febc62fca0bb00b858ae3cd9
master results: https://gist.github.com/li-roy/63b18e866e77b916c52ea12a23511de7

This pr is only slower on cases where the tensor is integral and the scalar is floating. This is expected because we need to convert the tensors to floating types in order to compare them properly. I couldn't think of a faster way to do this.

Still working on getting this to compile on windows.