@apaszke 2018-03-08
@li-roy couldn't we pre-process the scalar by taking the floor and using `<=` if `floor(x) != x` and `<` otherwise?