@li-roy 2018-03-09
@apaszke I think you mean `<` if `floor(x) != x` and `<=` otherwise? In any case that doesn't quite work because casting truncates towards 0. Thanks for the suggestion though, it gave me a very similar idea.

If `floor(x) != x`:
- for ge, use `ceil(x)` instead of `x`
- for le, use `floor(x)` instead of `x`
- for eq, just fill with `0`
- for ne, just fill with `1`

Ran the same benchmark script and it gives results that look more like master: https://gist.github.com/li-roy/164a2be5c341abccad8b534b2dce099e