@soumith 2018-03-09
Windows failure is

```
fatal error C1128: number of sections exceeded object file format limit: compile with /bigobj
```

Somehow your bigobj flag isn't kicking-in