@li-roy 2018-03-16
I did something similar to what we discussed, but I still couldn't get the type dispatch to work for both cpu and cuda. I don't think it's possible because cuda has the half type, which won't compile without cuda.

Also, I left the equals op in aten/src/THC/THCTensorMathCompareT.cuh for now because there's a function in  THCTensorMathPointwise that depends on it.