@ezyang 2018-03-23
@apaszke Can this go in even if the small comments are not addressed?