@apaszke 2018-03-23
The comments should be very easy to resolve, so I'd rather wait for a reply and get this in right afterwards.