@fmassa 2018-03-24
Quick question: do we need / want to remove the implementations in `TH` / THC` ? Up to now we have kept everything there, and just progressively started implementing natively on ATen, while also keeping the implementations where they were.