@li-roy 2018-03-28
Oh since the GPU perf test is back, this is failing again because it increases the time of cuda init significantly.